requirements.txt
contextual-client
python-dotenv
streamlit